feat: Refactor table parsers to convert twips and half points to points
- Updated `table_cell_properties_parser.py`:
  - Converted table cell width and margins from twips to points using `convert_twips_to_points`.

- Updated `table_grid_parser.py`:
  - Converted grid column widths from twips to points using `convert_twips_to_points`.

- Updated `table_properties_parser.py`:
  - Converted table width and indent from twips to points using `convert_twips_to_points`.
  - Converted table cell border sizes from half points to points using `convert_half_points_to_points`.
  - Converted table cell margins from twips to points using `convert_twips_to_points`.

- Updated `table_row_properties_parser.py`:
  - Converted row height from twips to points and ensured it is returned as a string.

- Updated Pydantic models in `table_models.py`:
  - Changed `width` fields in `TableWidth` and `TableIndent` models from `Optional[int]` to `Optional[float]`.
  - Changed margin values in `MarginProperties` model from `Optional[int]` to `Optional[float]`.
  - Changed `columns` field in `TableGrid` model from `List[int]` to `List[float]`.

These changes ensure consistent handling of measurement units by converting twips and half points to points, and allow Pydantic models to accept floating-point values where appropriate.

+++ docx_parsers/models/table_models.py
@@ -18,18 +18,18 @@ class ShadingProperties(BaseModel):
     color: Optional[str]
 
 class MarginProperties(BaseModel):
-    top: Optional[int]
-    left: Optional[int]
-    bottom: Optional[int]
-    right: Optional[int]
+    top: Optional[float]
+    left: Optional[float]
+    bottom: Optional[float]
+    right: Optional[float]
 
 class TableWidth(BaseModel):
     type: Optional[str]
-    width: Optional[int]
+    width: Optional[float]
 
 class TableIndent(BaseModel):
     type: Optional[str]
-    width: Optional[int]
+    width: Optional[float]
 
 class TableLook(BaseModel):
     firstRow: Optional[bool]
@@ -86,7 +86,7 @@ class TableProperties(BaseModel):
     tblLook: Optional[TableLook]
 
 class TableGrid(BaseModel):
-    columns: List[int]
+    columns: List[float]
 
 class Table(BaseModel):
     properties: Optional[TableProperties]
